EDMT-213 refactor : 기존에 잘못된 커밋 리팩토링
